requirements.txt
dotenv
google-generativeai
requests
pandas
tqdm
beautifulsoup4
selenium